Fixed a bug with deleting work history entries
This fixes a bug where the index for deleting a work history entry (from
the profile) wasn't carried through correctly, so that, ultimately,
`undefined` would be passed to `Array.prototype.splice` instead of a
Number. This meant that the user could, for instance, click the delete
button for the second work history entry, and the first one would be
deleted. Not good!

When fixing the bug I also noticed some unnecessary indirection that was
going on in the `<EmploymentForm />` and `<Educationform />` which I
cleaned up as well.

closes #2150
pr #2152
