+++ static/js/components/EducationForm.js
@@ -20,11 +20,6 @@ import ConfirmDeletion from './ConfirmDeletion';
 import SelectField from './inputs/SelectField';
 import CountrySelectField from './inputs/CountrySelectField';
 import StateSelectField from './inputs/StateSelectField';
-import {
-  openEditEducationForm,
-  openNewEducationForm,
-  deleteEducationEntry,
-} from '../util/profile_history_edit';
 import { educationEntriesByDate } from '../util/sorting';
 import { dialogActions } from './inputs/util';
 import {
@@ -46,6 +41,7 @@ import type {
 } from '../lib/validation/profile';
 import { formatMonthDate } from '../util/date';
 import FIELDS_OF_STUDY from '../data/fields_of_study';
+import { generateNewEducation } from "../util/util";
 
 const fieldOfStudyOptions = labelSort(_.map(FIELDS_OF_STUDY, (name, code) => ({
   value: code,
@@ -79,15 +75,45 @@ class EducationForm extends ProfileFormFields {
   };
 
   openEditEducationForm: Function = (index: number): void => {
-    openEditEducationForm.call(this, index);
+    const {
+      profile,
+      setEducationDialogIndex,
+      setEducationDegreeLevel,
+      setEducationDialogVisibility,
+    } = this.props;
+
+    let education = profile['education'][index];
+    setEducationDialogIndex(index);
+    setEducationDegreeLevel(education.degree_name);
+    setEducationDialogVisibility(true);
   };
 
-  openNewEducationForm: Function = (level: string, index: number): void => {
-    openNewEducationForm.call(this, level, index);
+  openNewEducationForm: Function = (level: string): void => {
+    const {
+      profile,
+      updateProfile,
+      setEducationDialogIndex,
+      setEducationDegreeLevel,
+      setEducationDialogVisibility,
+      validator,
+    } = this.props;
+    let clone = {
+      ...profile,
+      education: [...profile.education, generateNewEducation(level)]
+    };
+    updateProfile(clone, validator, true);
+    setEducationDialogIndex(clone.education.length - 1);
+    setEducationDegreeLevel(level);
+    setEducationDialogVisibility(true);
   };
 
-  deleteEducationEntry: Function = (): void => {
-    return deleteEducationEntry.call(this);
+  deleteEducationEntry: Function = (): Promise<*> => {
+    const { saveProfile, profile, ui } = this.props;
+    let clone = _.cloneDeep(profile);
+    if (ui.deletionIndex !== undefined && ui.deletionIndex !== null) {
+      clone['education'].splice(ui.deletionIndex, 1);
+    }
+    return saveProfile(educationValidation, clone, ui);
   };
 
   educationLevelRadioSwitch: Function = (level: Object): React$Element<*> => {
@@ -119,7 +145,7 @@ class EducationForm extends ProfileFormFields {
       if (level in educationLevelAnswers) {
         setEducationLevelAnswers(_.omit(educationLevelAnswers, [level]));
       }
-      this.openNewEducationForm(level, null);
+      this.openNewEducationForm(level);
     } else {
       setEducationLevelAnswers({
         ...educationLevelAnswers,
@@ -167,7 +193,7 @@ class EducationForm extends ProfileFormFields {
         <Cell col={12} className="profile-form-row add" key="add-row">
           <button
             className="mm-minor-action add-education-button"
-            onClick={() => this.openNewEducationForm(levelValue, null)}
+            onClick={() => this.openNewEducationForm(levelValue)}
           >
             Add degree
           </button>
